pyproject.toml
[project]
name = "ai-lounge-mcp"
version = "0.1.0"
description = "MCP server and CLI tool for posting to AI Lounge GitHub Discussions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ai-lounge-mcp = "ai_lounge_mcp.main:run"
ai-lounge = "ai_lounge_mcp.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
